requirements.txt
requests>=2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.31.0
torch==2.0.1
spacy==3.6.1 